feat: added PostgreSQL and Docker configuration

+++ src/main/java/com/nico/library/payload/response/BookResponse.java
@@ -6,14 +6,14 @@ import lombok.Data;
 @Data
 public class BookResponse
 {
-    private int id;
+    private Long id;
     private String title;
     private String author;
     private String plot;
     private String genre;
     private String ISBN;
 
-    public BookResponse(int id, String title, String author, String plot, String genre, String ISBN) {
+    public BookResponse(Long id, String title, String author, String plot, String genre, String ISBN) {
         this.id = id;
         this.title = title;
         this.author = author;
